pyproject.toml
[project]
name = "baes"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "jinja2>=3.1.6",
    "sqlalchemy>=2.0.0",
    "pyodbc>=4.0.39",
    "flask-cors>=4.0.0",
    "flask-migrate>=4.0.5",
    "flasgger>=0.9.7",
    "flask-login>=0.6.3",
    "flask-restx>=1.3.0",
    "paho-mqtt>=2.1.0",
    "requests>=2.32.4",
    "pyjwt>=2.8.0",
]
